+++ CHANGELOG.md
@@ -18,6 +18,9 @@
     * Add `POST /admin/v2/task-group/{task_group_id}/{deactivate|reactivate}` endpoints (\#2062, \#2078).
     * Remove `GET /auth/current-user/viewer-paths/` (\#2096).
     * Add `GET /auth/current-user/allowed-viewer-paths/`, with logic for `fractal-vizarr-viewer` authorization (\#2096).
+    * Add `POST /auth/group/{group_id}/add-user/{user_id}/` (\#2101).
+    * Add `POST /auth/group/{group_id}/remove-user/{user_id}/` (\#2101).
+    * Remove `new_user_ids` property from `PATCH /auth/group/{group_id}/` (\#2101).
     * Internals:
       * Fix bug in `_get_collection_task_group_activity_status_message` (\#2047).
       * Remove `valutc` validator for timestamps from API schemas, since it does not match with `psycopg3` behavior (\#2064).